fix(skills): consolidate to @/lib/ai/skills, delete stub dir (closes #211)
After Phase 2, the 3 AI surfaces (claude.ts, interview.ts,
transcript-analysis.ts) were importing from @/lib/skills (stubs PR #207
added as pre-reqs) instead of the canonical @/lib/ai/skills (real
helper from PR #208 + getHrSkillSettings from #206). The stub
loadHrSkill returned null unconditionally — meaning the HR_POLICY
block never reached Claude even when the tenant toggle was on.

Fix:
- Update 3 imports each in claude.ts, interview.ts, transcript-analysis.ts
- Delete src/lib/skills/ stub directory (index.ts + tenant-toggle.ts)
- Update vi.mock paths in skill-injection.test.ts to match

8/8 skill-injection tests pass against the new paths. The byte-identical
toggle-OFF guarantee still holds (test 1, 3, 5 assert system array
shape unchanged when settings.enabled = false).

This makes Phase 2's HR skill integration actually fire when tenant
admins enable it via /settings → AI Behaviour.
